+++ src/main/java/org/example/fishingconfig47/repositories/impl/RodRepositoryImpl.java
@@ -53,7 +53,7 @@ public class RodRepositoryImpl extends CustomCrudRepositoryImpl<Rod, Integer> im
     }
 
     @Override
-    public Rod findBudgetRod(double budget) {
+    public Rod findBudgetRod(float budget) {
         String jpql = "SELECT r FROM Rod r WHERE r.price <= :budget ORDER BY r.price DESC";
 
         TypedQuery<Rod> query = entityManager.createQuery(jpql, Rod.class);
@@ -64,7 +64,7 @@ public class RodRepositoryImpl extends CustomCrudRepositoryImpl<Rod, Integer> im
     }
 
     @Override
-    public List<Rod> findTop4ByPriceLessThanEqualOrderByPriceDesc(double maxPrice) {
+    public List<Rod> findTop4ByPriceLessThanEqualOrderByPriceDesc(float maxPrice) {
         String jpql = "SELECT r FROM Rod r WHERE r.price <= :maxPrice ORDER BY r.price DESC";
         return entityManager.createQuery(jpql, Rod.class)
                 .setParameter("maxPrice", maxPrice)
